rename: PopupDtoForWaitingResponse -> PopupSummaryResponse로 변경
