+++ backend/src/main/java/com/example/demo/application/dto/popup/PopupSummaryResponse.java
@@ -1,9 +1,7 @@
-package com.example.demo.application.dto.waiting;
-
-import com.example.demo.application.dto.popup.LocationResponse;
+package com.example.demo.application.dto.popup;
 
 // 다른 클래스에서 활용시 합의 필요
-public record PopupDtoForWaitingResponse(
+public record PopupSummaryResponse(
         Long popupId,
         String popupName,
         String popupImageUrl,
