+++ backend/src/main/java/com/example/demo/application/dto/waiting/WaitingResponse.java
@@ -1,5 +1,7 @@
 package com.example.demo.application.dto.waiting;
 
+import com.example.demo.application.dto.popup.PopupSummaryResponse;
+
 /**
  * 방문 내역 조회에서 사용할 Waiting 객체 응답 DTO
  */
@@ -10,5 +12,5 @@ public record WaitingResponse(
     String name, // 예약자 이름
     int peopleCount, // 예약 인원
     String contactEmail, // 예약 이메일
-    PopupDtoForWaitingResponse popup
+    PopupSummaryResponse popup
 ) {}
